# refactor(core): More efficient serde for ES modules in snapshot