Instead of relying on "serde_v8" which is very inefficient in serializing enums, I'm
hand rolling serde for "ModuleMap" data that is stored in the V8 snapshot to make
ES modules snapshottable.

Will post number comparison later, but I don't expect more than 5-10% improvement in
startup time.

EDIT: Benchmark results:
```
// this branch
Benchmark #2: ./target/release/deno run empty.js

  Time (mean ± σ):      21.4 ms ±   0.9 ms    [User: 15.6 ms, System: 6.4 ms]

  Range (min … max):    20.2 ms …  24.4 ms

// main branch
Benchmark #2: ./target/release/deno run empty.js

  Time (mean ± σ):      23.1 ms ±   1.2 ms    [User: 17.0 ms, System: 6.2 ms]

  Range (min … max):    21.0 ms …  26.0 ms

```